chore: Remove deprecated Pagination component from UI library and update references
