+++ components/ui/index.ts
@@ -34,8 +34,6 @@ export {
 
 export { default as Input } from "./input";
 export { default as ProgressBar } from "./progressbar";
-export { default as Pagination } from "./pagination";
-
 export { Toaster } from "./sonner";
 
 export { Select, SelectPrimitive } from "./select";
